Move BigQuery config class to separate module
