+++ python/lsst/dax/ppdb/bigquery/__init__.py
@@ -23,6 +23,7 @@ from .chunk_promoter import *
 from .chunk_uploader import *
 from .manifest import *
 from .ppdb_bigquery import *
+from .ppdb_bigquery_config import *
 from .ppdb_replica_chunk_extended import *
 from .query_runner import *
 from .sql_resource import *
